backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
langchain-aws>=0.1.0
langgraph>=0.1.0
boto3>=1.34.0
python-dotenv>=1.0.0
pydantic>=2.0.0
